pyproject.toml
[project]
name = "fastapi-learning"
version = "0.1.4"
description = "Learning FastAPI example: simple API exposing book endpoints with in-memory data, demonstrating path parameters and basic request/response behavior."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.17.2",
    "bcrypt==4.0.1",
    "fastapi>=0.124.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "passlib>=1.7.4",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.38.0",
]

[tool.pytest.ini_options]
testpaths = ["todoapp/tests"]

[tool.ruff.lint.extend-per-file-ignores]
"tests/*.py" = ["F401"]  # Ignore unused imports in all test files
